Update Readme for community edition
Airlock Microgateway 2.1 will be publicly available as community
edition. Therefore, the Docker Hub repo is public and without a
valid license the Airlock Microgateway works as community edition.
For further information consult the documentation:
https://docs.airlock.com/microgateway/latest/

Since the Airlock Minikube Example is designed for the premium
edition, the Readme has been updated to reflect that.